requirements.txt
python-telegram-bot==13.15
flask==2.3.3
pillow==10.3.0
